Finished the Grading Page, now finishing up the Leaderboard/Announcement page.
AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA

(aside from the fact that we're running out of time, everything is going well)

+++ app/src/main/java/com/risingstar/talentaachiva/domain/data/Submissions.kt
@@ -27,5 +27,5 @@ data class Judgement(
 @Parcelize
 data class Score(
     var criteria: Criteria? = null,
-    var amount : Double? = null
+    var amount: Int? = null
 ) : Parcelable
